Added complete domain support, examples and docs
