+++ Cargo.toml
@@ -16,5 +16,12 @@ license = "MIT"
 log = "0.4.0"
 glib-sys = { version = "0.9.1", features = ["v2_44"] }
 
+[features]
+redefine_log_macros = []
+
 [dev-dependencies]
 serial_test = "0.4.0"
+
+[[example]]
+name = "override_macro_domain"
+required-features = [ "redefine_log_macros" ]
